context_reset_demo/requirements.txt
openai
python-dotenv 